decouple kdb.js from elektrad
